Move reverse_documents into documents
